+++ src/file_management.rs
@@ -7,8 +7,7 @@ use std::fs::File;
 use std::io::{Read, Seek, SeekFrom};
 use std::{path::Path, time::Duration};
 
-use crate::utils;
-
+#[derive(Debug)]
 pub struct ObjectTags {
     pub managed_by: String,
     pub start_hash: String,
@@ -41,7 +40,9 @@ pub async fn presign_file(
     prefix: Option<String>,
     presigned_time: u64,
 ) -> Result<String, anyhow::Error> {
+    log::trace!("Presigning file: {:?}/{} in bucket {} for duration of {}", &prefix, &key, &bucket_name, &presigned_time);
     let expires_in = Duration::from_secs(presigned_time);
+    log::trace!("Sending get_object request that will presing the file");
     let presigned_request = client
         .get_object()
         .bucket(bucket_name)
@@ -74,15 +75,18 @@ pub struct PartialFileHash {
 }
 
 pub fn calculate_partial_hash(local_path: &Path) -> Result<PartialFileHash, anyhow::Error> {
+    log::trace!("Calculating partial hash of {:?}", &local_path);
     const SAMPLE_SIZE: usize = 8192;
 
     let mut file = File::open(local_path)?;
     let file_size = file.metadata()?.len();
+    log::trace!("File size of {:?} is {:?}", &local_path, &file_size);
 
     let mut start_buffer = vec![0; SAMPLE_SIZE];
     let start_bytes_read = file.read(&mut start_buffer)?;
     start_buffer.truncate(start_bytes_read);
     let start_hash = format!("{:x}", md5::compute(&start_buffer));
+    log::trace!("Start hash of {:?} is {:?}", &local_path, &start_hash);
 
     // This is just a check if the file is too small (less than 8KB as is the sample size by
     // default).
@@ -96,9 +100,10 @@ pub fn calculate_partial_hash(local_path: &Path) -> Result<PartialFileHash, anyh
         format!("{:x}", md5::compute(&end_buffer))
     } else {
         // The file is too small just use the start_hash
-        utils::print_warning("The file seems to be smaller than the sample size for hashing. Not sure how we got here.");
+        log::warn!("The file seems to be smaller than the sample size for hashing. Not sure how we got here.");
         start_hash.clone()
     };
+    log::trace!("End hash of {:?} is {:?}", &local_path, &end_hash);
 
     Ok(PartialFileHash {
         start_hash,
@@ -112,17 +117,30 @@ pub async fn file_exists_in_s3(
     bucket: &str,
     key: &str,
 ) -> Result<bool, anyhow::Error> {
+    log::trace!("Testing if {:?} exists in bucket {:?}", &key, &bucket);
     match client.head_object().bucket(bucket).key(key).send().await {
-        Ok(_) => Ok(true),
+        Ok(_) => {
+            log::trace!("File {:?} has been found in bucket {:?}", &key, &bucket);
+            Ok(true)
+        },
         Err(err) => match err {
             SdkError::ServiceError(err) => {
                 match err.err() {
                     // If the error NotFound is returned - return false
-                    HeadObjectError::NotFound(_) => Ok(false),
-                    other_err => Err(anyhow::anyhow!("S3 service error: {:?}", other_err)),
+                    HeadObjectError::NotFound(_) => {
+                        log::trace!("File {:?} not found in bucket {:?}", &key, &bucket);
+                        Ok(false)
+                    },
+                    other_err => {
+                        log::warn!("There was a service error when checking for file {:?} in bucket {:?}", &key, &bucket);
+                        Err(anyhow::anyhow!("S3 service error: {:?}", other_err))
+                    },
                 }
             }
-            other_err => Err(anyhow::anyhow!("S3 SDK error: {:?}", other_err)),
+            other_err => {
+                log::warn!("There was a SDK when checking for file {:?} in bucket {:?}", &key, &bucket);
+                Err(anyhow::anyhow!("S3 SDK error: {:?}", other_err))
+            },
         },
     }
 }
@@ -133,14 +151,27 @@ async fn get_file_metadata(
     bucket: &str,
     key: &str,
 ) -> Result<Option<aws_sdk_s3::operation::head_object::HeadObjectOutput>, anyhow::Error> {
+    log::trace!("Getting file metadata for {}:{}", &bucket, &key);
     match client.head_object().bucket(bucket).key(key).send().await {
-        Ok(output) => Ok(Some(output)),
+        Ok(output) => {
+            log::trace!("Metadata has been extracted for {:?} in bucket {:?}", &key, &bucket);
+            Ok(Some(output))
+        },
         Err(err) => match err {
             SdkError::ServiceError(err) => match err.err() {
-                HeadObjectError::NotFound(_) => Ok(None),
-                other_err => Err(anyhow::anyhow!("S3 service error: {:?}", other_err)),
+                HeadObjectError::NotFound(_) => {
+                    log::warn!("Unable to find {:?} in bucket {:?}, there seems to be an issue getting metadata", &key, &bucket);
+                    Ok(None)
+                },
+                other_err => {
+                    log::warn!("There was a service error when gettinf metadata for file {:?} in bucket {:?}", &key, &bucket);
+                    Err(anyhow::anyhow!("S3 service error: {:?}", other_err))
+                },
             },
-            other_err => Err(anyhow::anyhow!("S3 SDK error: {:?}", other_err)),
+            other_err => {
+                log::warn!("There was a S3 SDK error when gettinf metadata for file {:?} in bucket {:?}", &key, &bucket);
+                Err(anyhow::anyhow!("S3 SDK error: {:?}", other_err))
+            },
         },
     }
 }
@@ -152,6 +183,7 @@ async fn get_file_tags(
     key: &str,
 ) -> Result<Option<aws_sdk_s3::operation::get_object_tagging::GetObjectTaggingOutput>, anyhow::Error>
 {
+    log::trace!("Getting file tags for {}:{}", &bucket, &key);
     match client
         .get_object_tagging()
         .bucket(bucket)
@@ -159,8 +191,14 @@ async fn get_file_tags(
         .send()
         .await
     {
-        Ok(output) => Ok(Some(output)),
-        Err(err) => Err(anyhow::anyhow!("S3 service error: {:?}", err)),
+        Ok(output) => {
+            log::trace!("Tags for {} in {} were succesfully retrieved.", &key, &bucket);
+            Ok(Some(output))
+        },
+        Err(err) => {
+            log::warn!("There was a S3 Service error when getting tags for {} in {}.", &key, &bucket);
+            Err(anyhow::anyhow!("S3 service error: {:?}", err))
+        },
     }
 }
 
@@ -171,11 +209,16 @@ pub async fn quick_compare(
     local_object_tags: &ObjectTags,
     c: &Client,
 ) -> Result<bool, anyhow::Error> {
+    log::trace!("Comparing local and remote files: {:?} and {}/{}", &local_path, &bucket_name, &key);
     // Get file metadata
     let file = File::open(local_path)?;
     let file_size = file.metadata()?.len();
+    log::trace!("Local file {:?} size: {:?}", &local_path, &file_size);
+
     let object_metadata = get_file_metadata(c, bucket_name, key).await?;
+    log::trace!("Remote file {}{} metadata: {:#?}", &bucket_name, &key, &object_metadata);
     let object_tags = get_file_tags(c, bucket_name, key).await?;
+    log::trace!("Remote file {}{} tags: {:#?}", &bucket_name, &key, &object_metadata);
 
     // NOTE: Very complex way of making sure the length of my remote file is extracted
     // if I cannot do it, I just return 0 and we reupload
@@ -198,6 +241,7 @@ pub async fn quick_compare(
             },
         },
     };
+    log::trace!("The size of the remote file {}{}: {}", &bucket_name, &key, &s3_object_len);
 
     // Extracting the hash tags
     // FIX: Clean it up
@@ -208,6 +252,7 @@ pub async fn quick_compare(
         .find(|tag| tag.key() == "start_hash")
         .map(|tag| tag.value())
         .unwrap_or_default();
+    log::trace!("Remote file {}{} start_hash: {}", &bucket_name, &key, &remote_start_hash);
 
     let remote_end_hash = tags
         .tag_set()
@@ -215,6 +260,7 @@ pub async fn quick_compare(
         .find(|tag| tag.key() == "end_hash")
         .map(|tag| tag.value())
         .unwrap_or_default();
+    log::trace!("Remote file {}{} end_hash: {}", &bucket_name, &key, &remote_end_hash);
 
     // Compare the file size
     if file_size == s3_object_len {
@@ -224,12 +270,15 @@ pub async fn quick_compare(
             && local_object_tags.end_hash == remote_end_hash
         {
             //   If the same - presign
+            log::trace!("Both file are the same: local_object_tags.start_hash = {} == remote_start_hash {}; local_object_tags.end_hash = {} == remote_end_hash = {} ", &local_object_tags.start_hash, &remote_start_hash, &local_object_tags.end_hash, &remote_end_hash);
             Ok(true)
         } else {
+            log::trace!("The filenames are the same, but their partial hashes differ: local_object_tags.start_hash = {} != remote_start_hash {}; local_object_tags.end_hash = {} != remote_end_hash = {} ", &local_object_tags.start_hash, &remote_start_hash, &local_object_tags.end_hash, &remote_end_hash);
             println!("⚠️ | There seems to be a file with the same filename already at the destination. They are, also, the same sizes. HOWEVER, their partial hashes differ. I will assume that that they are different, so I will upload this one");
             Ok(false)
         }
     } else {
+        log::trace!("The filenames are the same, but their sizes differ: file_size {} != s3_object_len {}", &file_size, &s3_object_len);
         println!("⚠️ | There seems to be a file with the same filename already at the destination. They differ in sizes, I will assume that that they are different, so I will upload this one");
         Ok(false)
     }